+++ LapIt/ViewModels/HistoryViewModel.swift
@@ -24,8 +24,8 @@ class HistoryViewModel: ObservableObject {
         var user_email: String
         var firstName: String
         var secondName: String
-        var km: Int
-        var min: Int
+        var km: Float
+        var min: Float
     }
     
     @Published public var firstName = ""
@@ -91,8 +91,8 @@ class HistoryViewModel: ObservableObject {
                         let user_email = value["user_email"] as? String ?? ""
                         let firstName = try await network.getUserFirstName(email: user_email)
                         let secondName = try await network.getUserSecondName(email: user_email)
-                        let km = value["km"] as? Int ?? 0
-                        let min = value["min"] as? Int ?? 0
+                        let km = value["km"] as? Float ?? 0
+                        let min = value["min"] as? Float ?? 0
                         let user = User(user_email: user_email, firstName: firstName, secondName: secondName, km: km, min: min)
                         users.append(user)
                 }
